+++ R/tidync.R
@@ -1,0 +1,5 @@
+#' @name nc_grids 
+#' @export 
+nc_grids.tidync <- function(x, ...) {
+  x[["grids"]]
+}